requirements.txt
absl-py==0.14.1
altair==4.2.0
apache-beam==2.37.0
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.5.0
astunparse==1.6.3
attrs==21.2.0
audioread==3.0.0
avro-python3==1.10.2
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.5
cachetools==4.2.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
clang==5.0
click==8.0.4
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
contextlib2==21.6.0
crcmod==1.7
cycler==0.10.0
Cython==0.29.28
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
dill==0.3.1.1
dj-database-url==0.5.0
Django==4.0.2
django-environ==0.8.1
django-heroku==0.3.1
dm-tree==0.1.6
docopt==0.6.2
entrypoints==0.3
environ==1.0
fastavro==1.4.10
Flask==2.0.3
Flask-Cors==3.0.10
flat-api==0.8.1
flatbuffers==1.12
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.10
grpcio==1.41.0
gunicorn==20.1.0
h5py==3.1.0
hdfs==2.6.0
httplib2==0.19.1
idna==3.2
importlib-metadata==5.1.0
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonpickle==2.1.0
jsonschema==4.0.1
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
librosa==0.9.2
llvmlite==0.39.1
lvis==0.5.3
lxml==4.8.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
midi2audio==0.1.1
mistune==0.8.4
more-itertools==8.12.0
music21==7.1.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
numba==0.56.4
numpy==1.21.5
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
opt-einsum==3.3.0
orjson==3.6.7
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
pooch==1.6.0
psycopg2==2.9.3
py-cpuinfo==8.0.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.8.0
pydot==1.4.2
Pygments==2.10.0
pymongo==3.12.3
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-slugify==6.1.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==301
pywinpty==1.1.4
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2022.3.15
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.4.2
rich==12.6.0
rsa==4.7.2
sacrebleu==2.0.0
scikit-learn==1.0.2
scipy==1.7.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
seqeval==1.2.2
six==1.15.0
smmap==5.0.0
soundfile==0.11.0
soupsieve==2.3.1
sqlparse==0.4.2
streamlit==1.15.2
tabulate==0.8.9
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.8.0
tensorflow-addons==0.16.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-io==0.24.0
tensorflow-io-gcs-filesystem==0.24.0
tensorflow-metadata==1.7.0
tensorflow-model-optimization==0.7.2
tensorflow-text==2.8.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
tf-models-official==2.8.0
tf-slim==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.1
tqdm==4.63.0
traitlets==5.1.0
typeguard==2.13.3
typing_extensions==4.4.0
tzdata==2021.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.7
validators==0.20.0
watchdog==2.2.0
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==2.0.2
whitenoise==6.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.11.0
